Remove notice of June 1st, 2023 update
The Privacy Policy went into effect as scheduled on June 1st, 2023. 

Regrettably, this section was not removed in step with the update. The section linking to the previous Privacy Policy should have been removed in particular, as it (now incorrectly) identifies the superseded Policy as “current”.

As this edit does not affect the Privacy Policy in its substance, this Pull Request will only be kept open for 7 Days (until 2023-07-21 at 00:00 UTC) and not reflected in the “Last updated” and “Effective” dates.